Some reorganization in sources in order to correct writer functionality support.
