+++ src/librepgp/Makefile.am
@@ -10,5 +10,6 @@ librepgp_la_SOURCES     =  \
     reader.c \
     repgp.c \
     validate.c \
+    stream-common.c \
     stream-parse.c \
     stream-write.c
